Kitchen components added

+++ Backend/Services/AuthService.cs
@@ -24,47 +24,57 @@ namespace Backend.Services
            .FirstOrDefaultAsync(u => u.Email == loginRequest.Email && u.Password == loginRequest.Password);
 
             if (user == null)
-                return null;
+                throw new  UnauthorizedAccessException("Invalid email or password");
+
+            var session = await _dbContext.UserSession
+                 .FirstOrDefaultAsync(s => s.UserId == user.Id);
+
+            if (session != null)
+            {
+                if (session.EndTime <= DateTime.Now) _dbContext.UserSession.Remove(session);
+
+                else throw new ArgumentException("User already logged in");
+            }
 
             string token = $"{Guid.NewGuid()}_{user.UserType}";
 
-            var session = new UserSession
+            var newSession = new UserSession
             {
                 Token = token,
                 UserType = user.UserType,
                 UserId = user.Id,
                 StartTime = DateTime.Now,
-                EndTime = DateTime.Now.AddHours(1)
+                EndTime = DateTime.Now.AddHours(5)
             };
 
-            _dbContext.UserSession.Add(session);
+            _dbContext.UserSession.Add(newSession);
             await _dbContext.SaveChangesAsync();
 
             return new UserSessionDTO
             {
                 Token = token,
                 UserType = user.UserType.ToString(),
-                StartTime = session.StartTime,
-                EndTime = session.EndTime
+                StartTime = newSession.StartTime,
+                EndTime = newSession.EndTime
             };
         }
 
         public async Task<UserSession?> ValidateTokenAsync(string token)
         {
-            
-        var session = await _dbContext.UserSession.FirstOrDefaultAsync(s => s.Token == token);
 
-        if (session == null)
-            return null;
+            var session = await _dbContext.UserSession.FirstOrDefaultAsync(s => s.Token == token);
 
-        if (session.EndTime <= DateTime.Now)
-        {
-            _dbContext.UserSession.Remove(session);
-            await _dbContext.SaveChangesAsync();
-            return null;
-        }
+            if (session == null)
+                return null;
 
-        return session;
+            if (session.EndTime <= DateTime.Now)
+            {
+                _dbContext.UserSession.Remove(session);
+                await _dbContext.SaveChangesAsync();
+                return null;
+            }
+
+            return session;
         }
 
     }
